Ensure norm succeeds on numerical errors in inner
Due to numerical errors the inner product can sometimes return an infinitesimal negative number (e.g. -1e-26). This negative number causes the sqrt to fail with a domain error

This fix ensures that the result of the inner product of X and X is always greater than or equal to zero (which it should be by definition)